.NET 5 was a failed experiment
The .NET Core Windows Forms designer is still almost completely broken.  And a built .exe requires a whole folder full of several megabytes of DLL files instead of the (essentially) standalone .exe you get from a 4.7.2 Framework build.

This technically undoes #5, but I think I'm going to leave it closed for now.  Maybe in a few more years this won't be as terrible an experience.